[코드리뷰 반영] 삭제전 상태 체크 및 QuestionRemoveValidator inner class로 변경
