+++ src/main/java/nextstep/qna/domain/Question.java
@@ -1,7 +1,7 @@
 package nextstep.qna.domain;
 
+import nextstep.qna.CannotDeleteException;
 import nextstep.qna.domain.generator.SimpleIdGenerator;
-import nextstep.qna.domain.validate.QuestionRemoveValidator;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
@@ -128,5 +128,38 @@ public class Question {
     @Override
     public int hashCode() {
         return Objects.hash(id, writer, answers, title, contents);
+    }
+
+    private static class QuestionRemoveValidator {
+
+        public static void validate(Question question, NsUser requestUser) {
+            validateAlreadyDeleted(question);
+            validateAuthorization(question, requestUser);
+            validateAboutAnswer(question);
+        }
+
+        private static void validateAlreadyDeleted(Question question) {
+            if (question.isDeleted()) {
+                throw new CannotDeleteException("이미 질문이에요 :(");
+            }
+        }
+
+        private static void validateAuthorization(Question question, NsUser requestUser) throws CannotDeleteException {
+            if (!question.isOwner(requestUser)) {
+                throw new CannotDeleteException("글 작성자만 삭제 가능해요 :( (요청 사용자 : " + requestUser.getUserId() + ")");
+            }
+        }
+
+        private static void validateAboutAnswer(Question question) {
+
+            if (!question.hasAnswer()) {
+                return;
+            }
+
+            if (question.hasAnotherOwner()) {
+                throw new CannotDeleteException("다른분이 작성한 답변글이 존재해서 삭제 불가능 해요 :(");
+            }
+        }
+
     }
 }
